pyproject.toml
[tool.poetry]
name = "drf-playground"
version = "0.1.0"
description = ""
authors = ["Guillem MC <guimarca@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
djangorestframework = "^3.13.1"
django-filter = "^21.1"
requests = "^2.26.0"
pytest-django = "^4.5.2"
testcontainers = "^3.5.0"
psycopg2-binary = "^2.9.3"

[tool.poetry.extras]
testcontainers = ["postgres"]

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.12b0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "drf_playground.settings"
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
    "integration",
]